refactor: change file location and remove unused dependencies
