+++ smeem-common/build.gradle
@@ -1,7 +1,6 @@
 project(':smeem-common') {
     dependencies {
         implementation 'org.springframework.boot:spring-boot-starter-web'
-        implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     }
 
     tasks.bootJar {
